sdk/python/examples/requirements.txt
# Example dependencies
# Install with: pip install -r examples/requirements.txt

# SDK (install from parent directory)
# pip install -e ..

# Example-specific dependencies
openai>=1.0.0
python-dotenv>=1.0.0
